ajustando retornos do controllers e tipos de atributos em modelos e dtos

+++ src/main/java/com/amandaLinger/notasFiscais/dto/ProdutoDto.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
@@ -17,11 +19,11 @@ public class ProdutoDto {
     private String nome;
 
     @NotNull
-    private Double preco;
+    private BigDecimal preco;
 
     private String descricao;
 
     @NotNull
-    private Integer quantidade;
+    private BigDecimal quantidade;
 
 }
